Removed everything not related to the engine
