+++ .gitignore
@@ -6,3 +6,4 @@
 hs_err_pid*
 /.idea/
 /TextEngine.iml
+out/
