+++ src/com/jhjavadev/textengine/game/TestStage.java
@@ -1,0 +1,22 @@
+package com.jhjavadev.textengine.game;
+
+import com.jhjavadev.textengine.console.TextConsole;
+
+public class TestStage extends Stage {
+
+	public void printStart(TextConsole c) {
+
+	}
+
+	public boolean update(String cmd, String[] args, TextConsole c) {
+		return false;
+	}
+
+	public void printHelp(TextConsole c) {
+
+	}
+
+	public void reset() {
+
+	}
+}
